# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The user has been successfully registered.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request."
          }
        },
        "summary": "Register a new user",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The user has been successfully logged in.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          }
        },
        "summary": "Login an existing user",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "AuthController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Return the current user profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user profile",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/groups/me": {
      "get": {
        "operationId": "GroupController_getMyGroups",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserGroupMembershipDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Liste des groupes de l'utilisateur connecté",
        "tags": [
          "groups"
        ]
      }
    },
    "/groups/search": {
      "get": {
        "operationId": "GroupController_searchGroups",
        "parameters": [
          {
            "name": "name",
            "required": false,
            "in": "query",
            "schema": {
              "example": "Warriors",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Groupes correspondants"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Rechercher un groupe public par nom (vide = suggestions)",
        "tags": [
          "groups"
        ]
      }
    },
    "/groups/join/{code}": {
      "post": {
        "operationId": "GroupController_joinGroupByCode",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Groupe rejoint"
          },
          "404": {
            "description": "Code invalide"
          },
          "409": {
            "description": "Déjà membre"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Rejoindre un groupe privé via code d'invitation",
        "tags": [
          "groups"
        ]
      }
    },
    "/groups": {
      "post": {
        "operationId": "GroupController_createGroup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserGroupMembershipDto"
                }
              }
            }
          },
          "403": {
            "description": "Niveau insuffisant"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Créer un groupe (niveau ≥ 3 requis)",
        "tags": [
          "groups"
        ]
      }
    },
    "/groups/{id}": {
      "get": {
        "operationId": "GroupController_getGroupDetails",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Détails du groupe"
          },
          "404": {
            "description": "Groupe non trouvé"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Détails du groupe (membres, win streak, défi en cours)",
        "tags": [
          "groups"
        ]
      },
      "delete": {
        "operationId": "GroupController_deleteGroup",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Groupe supprimé"
          },
          "403": {
            "description": "Accès refusé"
          },
          "404": {
            "description": "Groupe non trouvé"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Dissoudre le groupe (admin uniquement)",
        "tags": [
          "groups"
        ]
      }
    },
    "/groups/{id}/join": {
      "post": {
        "operationId": "GroupController_joinGroupById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Groupe rejoint"
          },
          "403": {
            "description": "Groupe privé"
          },
          "404": {
            "description": "Groupe non trouvé"
          },
          "409": {
            "description": "Déjà membre"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Rejoindre un groupe public",
        "tags": [
          "groups"
        ]
      }
    },
    "/groups/{id}/members/{userId}": {
      "delete": {
        "operationId": "GroupController_removeMember",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Membre exclu"
          },
          "403": {
            "description": "Accès refusé"
          },
          "404": {
            "description": "Groupe non trouvé"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Exclure un membre (admin uniquement)",
        "tags": [
          "groups"
        ]
      }
    },
    "/groups/{id}/challenge": {
      "get": {
        "operationId": "GroupController_getGroupChallenge",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "groupChallengeId": "uuid",
                    "challenge": {
                      "id": "uuid",
                      "title": "...",
                      "description": "..."
                    },
                    "weekStartAt": "2026-05-11T00:00:00.000Z",
                    "weekEndAt": "2026-05-17T23:59:59.999Z",
                    "completedCount": 3,
                    "totalMembers": 5,
                    "progressRatio": 0.6
                  }
                }
              }
            }
          },
          "404": {
            "description": "Aucun défi cette semaine"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Défi de la semaine + barre de progression",
        "tags": [
          "groups"
        ]
      }
    },
    "/groups/{id}/challenge/complete": {
      "post": {
        "operationId": "GroupController_completeChallenge",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Défi complété"
          },
          "404": {
            "description": "Aucun défi cette semaine"
          },
          "409": {
            "description": "Défi déjà complété"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Marquer le défi hebdomadaire comme complété",
        "tags": [
          "groups"
        ]
      }
    },
...
  },
  "info": {
    "title": "Projet-AGL API",
    "description": "The API documentation for Projet-AGL backend",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "The email of the user",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "description": "The password of the user (min 6 characters)",
            "example": "password123",
            "minLength": 6
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "dc147d80-0ee4-42db-84f0-4150fa0593f3"
          },
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "email_confirmed_at": {
            "type": "string"
          },
          "last_sign_in_at": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "email"
        ]
      },
      "SessionDto": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string"
          },
          "token_type": {
            "type": "string",
            "example": "bearer"
          },
          "expires_in": {
            "type": "number",
            "example": 3600
          },
          "expires_at": {
            "type": "number"
          }
        },
        "required": [
          "access_token",
          "token_type",
          "expires_in",
          "expires_at"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserDto"
          },
          "session": {
            "$ref": "#/components/schemas/SessionDto"
          }
        },
        "required": [
          "user"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "The email of the user",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "description": "The password of the user",
            "example": "password123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "ParcoursDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "de879435-0ee4-42db-84f0-4150fa0593f3"
          },
          "slug": {
            "type": "string",
            "example": "decouverte"
          },
          "name": {
            "type": "string",
            "example": "DÉCOUVERTE"
          },
          "description": {
            "type": "string",
            "example": "L’essentiel rapidement. Comprendre mon score sans surcharge avec des défis simples."
          },
          "defis_per_period": {
            "type": "number",
            "example": 1
          },
          "quizz_per_period": {
            "type": "number",
            "example": 1
          },
          "period_type": {
            "type": "string",
            "example": "WEEK"
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "description",
          "defis_per_period",
          "quizz_per_period",
          "period_type"
        ]
      },
      "UserProfileResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "dc147d80-0ee4-42db-84f0-4150fa0593f3"
          },
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "firstName": {
            "type": "object",
            "nullable": true,
            "example": "Jean"
          },
          "lastName": {
            "type": "object",
            "nullable": true,
            "example": "Dupont"
          },
          "feuilles": {
            "type": "number",
            "example": 450
          },
          "niveau": {
            "type": "number",
            "example": 5
          },
          "onboardingCompleted": {
            "type": "boolean",
            "example": true
          },
          "hasOnboardingBilan": {
            "type": "boolean",
            "description": "Au moins un bilan carbone enregistré en base (ex. quiz d’onboarding soumis), indépendamment du choix de parcours.",
            "example": true
          },
          "parcours": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ParcoursDto"
              }
            ]
          },
          "lastSimpleChallengeCompletedAt": {
            "type": "string",
            "nullable": true,
            "description": "Horodatage ISO de la dernière complétion du défi Accueil (limite une fois par jour UTC).",
            "example": "2026-05-12T08:30:00.000Z"
          }
        },
        "required": [
          "id",
          "email",
          "feuilles",
          "niveau",
          "onboardingCompleted",
          "hasOnboardingBilan"
        ]
      },
...
      "GroupSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "member_count": {
            "type": "number"
          },
          "is_private": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "member_count",
          "is_private",
          "created_at",
          "updated_at"
        ]
      },
      "UserGroupMembershipDto": {
        "type": "object",
        "properties": {
          "community": {
            "$ref": "#/components/schemas/GroupSummaryDto"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "MEMBER"
            ]
          },
          "joined_at": {
            "type": "string"
          },
          "has_pending_defi": {
            "type": "boolean"
          }
        },
        "required": [
          "community",
          "role",
          "joined_at",
          "has_pending_defi"
        ]
      },
      "CreateGroupDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Les Éco-Warriors"
          },
          "description": {
            "type": "string",
            "example": "Groupe dédié aux défis écologiques"
          },
          "is_public": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "name",
          "is_public"
        ]
      },
...
    }
  }
}